Move ui controller to a seperate file
